+++ src/components/button/Button.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import classNames from 'classnames'
 import styles from './Button.module.scss'
