+++ src/components/icon/Icon.tsx
@@ -1,15 +1,21 @@
 import Image from 'next/image'
 
-import showPassword from '@/assets/eye-open.svg'
-import hidePassword from '@/assets/eye-closed.svg'
-
 interface IconProps {
   type: 'show' | 'hide'
   alt?: string
   [x: string]: any
 }
 
-export default function Icon({ type, alt = '', ...restProps }: IconProps) {
+const showPassword = '/auth/eye-open.svg'
+const hidePassword = '/auth/eye-closed.svg'
+
+export default function Icon({
+  type,
+  alt = '',
+  width = 24,
+  height = 24,
+  ...restProps
+}: IconProps) {
   let src = ''
 
   switch (type) {
@@ -23,5 +29,7 @@ export default function Icon({ type, alt = '', ...restProps }: IconProps) {
       throw new Error('지원하는 아이콘 타입이 존재하지 않습니다.')
   }
 
-  return <Image src={src} alt={alt} {...restProps} />
+  return (
+    <Image src={src} alt={alt} width={width} height={height} {...restProps} />
+  )
 }
