+++ src/components/icon/index.ts
@@ -1,0 +1,3 @@
+import Icon from './Icon'
+
+export { Icon }
